requirements.txt
scipy
pillow
pandas
click